feat(cli): move tokenization runtime loading into config loader
